+++ .gitignore
@@ -5,3 +5,5 @@
 /bin/tacentview
 /bin/pvr
 /out/
+
+/logfile.txt
